Crate gccjit

Structs

Block

CompileResult
Represents a successful compilation of a context. This type provides the means to access compiled functions and globals. JIT compiled functions are exposted to Rust as an extern "C" function pointer.

Context
Wrapper around a GCC JIT context object that keeps the state of the JIT compiler. In GCCJIT, this object is responsible for all memory management of JIT data structures, and as such anything made from this context must have a lifetime strictly less than this object.

Field

Function

LValue

Location

Object

Parameter

RValue

Struct

Type
A representation of a type, as it is known to the JIT compiler. Types can be created through the Typeable trait or they can be created dynamically by composing Field types.

Enums

BinaryOp

Concrete

FunctionType

Opaque

OptimizationLevel
Represents an optimization level that the JIT compiler will use when compiling your code.

UnaryOp

Traits

ToLValue

ToObject

ToRValue

Typeable