Enum gccjit::FunctionType

pub enum FunctionType {
    Exported,
    Internal,
    Extern,
    AlwaysInline,
}

FunctionType informs gccjit what sort of function a new function will be. An exported function is a function that will be exported using the CompileResult interface, able to be called outside of the jit. An internal function is a function that cannot be called outside of jitted code. An extern function is a function with external linkage, and always inline is a function that is always inlined wherever it is called and cannot be accessed outside of the jit.

Variants

Exported

Defines a function that is "exported" by the JIT and can be called from Rust.

Internal

Defines a function that is internal to the JIT and cannot be called from Rust, but can be called from jitted code.

Extern

Defines a function with external linkage.

AlwaysInline

Defines a function that should always be inlined whenever it is called. A function with this type cannot be called from Rust. If the optimization level is None, this function will not actually be inlined, but it still can only be called from within jitted code.