Struct gccjit::Context

pub struct Context<'ctx> {
    // some fields omitted
}

Wrapper around a GCC JIT context object that keeps the state of the JIT compiler. In GCCJIT, this object is responsible for all memory management of JIT data structures, and as such anything made from this context must have a lifetime strictly less than this object.

It's possible to create a child context from a parent context. In that case, the child context must have a lifetime strictly less than the parent context.

Methods

impl<'ctx> Context<'ctx>

fn set_program_name(&self, name: &str)

Sets the program name reported by the JIT.

fn set_optimization_level(&self, level: OptimizationLevel)

Sets the optimization level that the JIT compiler will use. The higher the optimization level, the longer compilation will take.

fn compile(&self) -> CompileResult

Compiles the context and returns a CompileResult that contains the means to access functions and globals that have currently been JIT compiled.

fn new_child_context<'b>(&'b self) -> Context<'b>

Creates a new child context from this context. The child context is a fully-featured context, but it has a lifetime that is strictly less than the lifetime that spawned it.

fn new_location<'a>(&'a self, filename: &str, line: i32, col: i32) -> Location<'a>

Creates a new location for use by gdb when debugging a JIT compiled program. The filename, line, and col are used by gdb to "show" your source when in a debugger.

fn new_type<'a, T: Typeable>(&'a self) -> Type<'a>

Constructs a new type for any type that implements the Typeable trait. This library only provides a handful of implementations of Typeable for some primitive types - utilizers of this library are encouraged to provide their own types that implement Typeable for ease of type creation.

fn new_field<'a>(&'a self, loc: Option<Location<'a>>, ty: Type<'a>, name: &str) -> Field<'a>

Constructs a new field with an optional source location, type, and name. This field can be used to compose unions or structs.

fn new_array_type<'a>(&'a self, loc: Option<Location<'a>>, ty: Type<'a>, num_elements: i32) -> Type<'a>

Constructs a new array type with a given base element type and a size.

fn new_struct_type<'a>(&'a self, loc: Option<Location<'a>>, name: &str, fields: &[Field<'a>]) -> Struct<'a>

Constructs a new struct type with the given name, optional source location, and a list of fields. The returned struct is concrete and new fields cannot be added to it.

fn new_opaque_struct_type<'a>(&'a self, loc: Option<Location<'a>>, name: &str) -> Struct<'a>

Constructs a new struct type whose fields are not known. Fields can be added to this struct later, but only once.

fn new_union_type<'a>(&'a self, loc: Option<Location<'a>>, name: &str, fields: &[Field<'a>]) -> Type<'a>

Creates a new union type from a set of fields.

fn new_function_pointer_type<'a>(&'a self, loc: Option<Location<'a>>, return_type: Type<'a>, param_types: &[Type<'a>], is_variadic: bool) -> Type<'a>

Creates a new function pointer type with the given return type parameter types, and an optional location. The last flag can make the function variadic, although Rust can't really handle the varargs calling convention.

fn new_function<'a>(&'a self, loc: Option<Location<'a>>, kind: FunctionType, return_ty: Type<'a>, params: &[Parameter<'a>], name: &str, is_variadic: bool) -> Function<'a>

Creates a new function with the given function kind, return type, parameters, name, and whether or not the function is variadic. It's not currently possible to call variadic functions from Rust right now, so that option is turned off for now.

fn new_binary_op<'a>(&'a self, loc: Option<Location<'a>>, op: BinaryOp, ty: Type<'a>, left: RValue<'a>, right: RValue<'a>) -> RValue<'a>

Creates a new binary operation between two RValues and produces a new RValue.

fn new_unary_op<'a>(&'a self, loc: Option<Location<'a>>, op: UnaryOp, ty: Type<'a>, rvalue: RValue<'a>) -> RValue<'a>

Creates a unary operation on one RValue and produces a result RValue.

fn new_call<'a>(&'a self, loc: Option<Location<'a>>, func: Function<'a>, args: &[RValue<'a>]) -> RValue<'a>

Creates a function call to a function object with a given number of parameters. The RValue that is returned is the result of the function call.

fn new_call_through_ptr<'a>(&'a self, loc: Option<Location<'a>>, fun_ptr: RValue<'a>, args: &[RValue<'a>]) -> RValue<'a>

Creates an indirect function call that dereferences a function pointer and attempts to invoke it with the given arguments. The RValue that is returned is the result of the function call.

fn new_cast<'a>(&'a self, loc: Option<Location<'a>>, rvalue: RValue<'a>, dest_type: Type<'a>) -> RValue<'a>

Cast an RValue to a specific type. I don't know what happens when the cast fails yet.

fn new_array_access<'a>(&'a self, loc: Option<Location<'a>>, array_ptr: RValue<'a>, index: RValue<'a>) -> LValue<'a>

Creates an LValue from an array pointer and an offset. The LValue can be the target of an assignment, or it can be converted into an RValue (i.e. loaded).

fn new_rvalue_from_long<'a>(&'a self, ty: Type<'a>, value: i64) -> RValue<'a>

Creates a new RValue from a given long value.

fn new_rvalue_from_int<'a>(&'a self, ty: Type<'a>, value: i32) -> RValue<'a>

Creates a new RValue from a given int value.

fn new_rvalue_from_double<'a>(&'a self, ty: Type<'a>, value: f64) -> RValue<'a>

Creates a new RValue from a given double value.

fn new_rvalue_zero<'a>(&'a self, ty: Type<'a>) -> RValue<'a>

Creates a zero element for a given type.

fn new_rvalue_one<'a>(&'a self, ty: Type<'a>) -> RValue<'a>

Creates a one element for a given type.

unsafe fn new_rvalue_from_ptr<'a>(&'a self, ty: Type<'a>, value: *mut u8) -> RValue<'a>

Creates an RValue for a raw pointer.

fn new_null<'a>(&'a self, ty: Type<'a>) -> RValue<'a>

Creates a null RValue.

fn new_string_literal<'a>(&'a self, value: &str) -> RValue<'a>

Creates a string literal RValue.

fn dump_reproducer_to_file(&self, path: &str)

Dumps a small C file to the path that can be used to reproduce a series of API calls. You should only ever need to call this if you are debugging a segfault in gccjit or this library.

fn new_parameter<'a>(&'a self, loc: Option<Location<'a>>, ty: Type<'a>, name: &str) -> Parameter<'a>

Creates a new parameter with a given type, name, and location.