Struct gccjit::LValue

pub struct LValue<'ctx> {
    // some fields omitted
}

An LValue in gccjit represents a value that has a concrete location in memory. A LValue can be converted into an RValue through the ToRValue trait. It is also possible to get the address of an LValue.

Methods

impl<'ctx> LValue<'ctx>

fn access_field(&self, loc: Option<Location<'ctx>>, field: Field<'ctx>) -> LValue<'ctx>

Given an LValue x and a Field f, gets an LValue for the field access x.f.

fn get_address(&self, loc: Option<Location<'ctx>>) -> RValue<'ctx>

Given an LValue x, returns the RValue address of x, akin to C's &x.