Struct gccjit::RValue

pub struct RValue<'ctx> {
    // some fields omitted
}

An RValue is a value that may or may not have a storage address in gccjit. RValues can be dereferenced, used for field accesses, and are the parameters given to a majority of the gccjit API calls.

Methods

impl<'ctx> RValue<'ctx>

fn get_type(&self) -> Type<'ctx>

Gets the type of this RValue.

fn access_field(&self, loc: Option<Location<'ctx>>, field: Field<'ctx>) -> LValue<'ctx>

Given an RValue x and a Field f, returns an RValue representing C's x.f.

fn dereference_field(&self, loc: Option<Location<'ctx>>, field: Field<'ctx>) -> LValue<'ctx>

Given an RValue x and a Field f, returns an LValue representing C's x->f.

fn dereference(&self, loc: Option<Location<'ctx>>) -> LValue<'ctx>

Given a RValue x, returns an RValue that represents *x.