Enum gccjit::FunctionType

pub enum FunctionType {
    Exported,
    Internal,
    Extern,
    AlwaysInline,
}

FunctionType informs gccjit what sort of function a new function will be. An exported function is a function that will be exported using the CompileResult interface, able to be called outside of the jit. An internal function is a function that cannot be called outside of jitted code. An extern function is a function with external linkage, and always inline is a function that is always inlined wherever it is called and cannot be accessed outside of the jit.

Variants

Exported

Internal

Extern

AlwaysInline