Struct gccjit::CompileResult

pub struct CompileResult {
    // some fields omitted
}

Represents a successful compilation of a context. This type provides the means to access compiled functions and globals. JIT compiled functions are exposted to Rust as an extern "C" function pointer.

Methods

impl CompileResult

fn get_function<'a, A, R>(&'a self, name: &str) -> Option< extern "C" fn(A) -> R>

Gets a function pointer to a JIT compiled function. If the function does not exist (wasn't compiled by the Context that produced this CompileResult), this function returns None.

unsafe fn get_global<T>(&self, name: &str) -> *mut T

Who knows what this thing does yet.